Change simply public API to be an opaque struct
